Add UI test framework
Moved the UI code into a separate Python library.

* tox allows us test multiple Python interpreters
* pytest allows us to write tests with standard assert calls
* pytest-cov reports how much code coverage we're getting

Note: test is named uwhdui/ui_tests.py, not tests/ui.py to keep
the unit tests as close as possible to the source code they test.

Fixes #5
