+++ tools/uwhdui/.coveragerc
@@ -1,0 +1,2 @@
+[run]
+omit = *_test.py
