+++ tools/uwhdui/setup.py
@@ -1,0 +1,7 @@
+from setuptools import setup, find_packages
+
+setup(
+    name='uwhdr',
+    version='1.0.0',
+    packages=find_packages(),
+)
