+++ tools/uwhdui/tox.ini
@@ -1,0 +1,17 @@
+[tox]
+envlist = py27
+
+[testenv]
+deps =
+    pytest
+    pytest-cov
+
+commands = py.test
+
+[pytest]
+addopts =
+    --cov=uwhdui
+    --cov-report=term-missing
+
+norecursedirs =
+    .tox
