+++ tools/uwhdui/uwhdui/ui_test.py
@@ -1,0 +1,4 @@
+from . import ui
+
+def test_sized_frame():
+    assert ui.sized_frame(None, 1, 2)
